Revert "30939 Migrating CAs from App Config JSON to their own table (#31739)"
This reverts commit da55b5f75a02b5ce36f4faab9adf97a6fec7bc8c.
